[DPE-8403] Improve error handling for missing permissions on user secrets (#156)
## Description of issue or feature:
Users can configure a password for the root user with a user secret. If
this secret is not granted to the etcd application, the `secret_changed`
/ `config_changed` event for processing the secret will fail and a
status message gets displayed.

But after updating the secret permissions, no new event is triggered and
the desired password can not be applied, until users manually update the
config or secret.

[DPE-8403](https://warthogs.atlassian.net/browse/DPE-8403)

## Solution:
In order to work around this gap, the failed events with permission
errors should be deferred and retried later. This allows for processing
the password, once users have granted permissions for the secret to the
etcd application.

## Unrelated changes:
- fix an issue with status computation while TLS switchover is in
progress (the previous fix in #138) was not sufficient
- observe `storage_attached` events for all defined storages (following
MongoDB approach as discussed with @patriciareinoso)
- fix a pretty critical issue with IP address changes when client TLS is
enabled (discovered while testing, see [test
run](https://github.com/canonical/charmed-etcd-operator/actions/runs/18227789817/job/51913188907#step:13:4956))
- the `config_changed` event crashes when the operator runs the health
check after restarting
- this happens because the client TLS cert used by the operator is not
valid anymore (was issued for the old IP address)
- if the failed health check would be ignored, all subsequent actions
such as updating the member information with the new peer url will also
fail
- the only way out is to request new client certs before doing anything
else on the cluster, and until then, defer the IP change
(`config_changed` event)

## How was this change tested?
- [X] Manually
- [x] Unit tests
- [x] Integration tests

All manual tests where executed on a remote VM with LXD:
- The status computation fix was tested manually by enabling and
disabling peer TLS.
- The renewal of TLS certs was tested with a manual network interruption
+ IP address change for one unit, both with peer and client TLS enabled;
an integration test has been added for IP address change with client
certificates to increase test coverage
- The `storage_attached` events where tested manually by deploying a
cluster with storage constraints, removing units and the application and
re-attaching the detached storage.
- The secret permission issue was tested manually by deploying an
application with configured secret, but without secret access given to
the charm. After failing to process this, the access was granted and the
next `update_status` event was checked for the password update.
Integration test coverage was also added for this.

[DPE-8403]:
https://warthogs.atlassian.net/browse/DPE-8403?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ