pin ubuntu 22.04 image used for lxd containers (#169)
## Description of issue or feature:
There is an issue with the jammy image used by lxd which is causing
charms running on 22.04 (eg. s3-integrator) to be stuck in "waiting for
machine" state and breaking integration tests.

This is caused by a kernel issue related to snapd. For more details,
follow [here](https://bugs.launchpad.net/snapd/+bug/2127244).

## Solution:
- pin the `amd64` image to a working one
- disable the integration tests on arm for `backup_restore_s3` and
`restore_verification` (they are still run on amd)

## How was this change tested?
- [ ] Manually
- [ ] Unit tests
- [X] Integration tests