fix flakey unit test (#180)
## Description of issue or feature:
Since recently, the unit test for removal of a unit started to fail (see
[example1](https://github.com/canonical/charmed-etcd-operator/actions/runs/19054645842/job/54422242315)
or [example
2](https://github.com/canonical/charmed-etcd-operator/actions/runs/19099391500/job/54567324710?pr=167#step:4:683)).
The failure rate on CI was about 50%, but could not be reproduced
locally.

## Solution:
The unit test's state needs to be set up more accurately to correctly
test the expected workflow (10 out of 10 test runs succeeded).

## How was this change tested?
- [ ] Manually
- [X] Unit tests
- [ ] Integration tests