[DPE-8989] Bugfix: allow same TLS-provider as client application (#190)
This PR addresses #189 

## Description of issue or feature:
If charmed-etcd and a client application share the same TLS provider and
a CA-rotation happens, there are multiple issues because of the
interfering events:

1) Interfering access from client and etcd on the `mtls-cert` secret if
the client uses data-interfaces v0
This results in etcd running into `hook failed: "secret-changed"` with
the following error:
```
ops.model.ModelError: ERROR getting latest secret revision: secret label "etcd-client.10.2e72e7f4350be731.mtls.secret" for consumer "unit-charmed-etcd-0" already exists
```
This is possibly a data-interfaces v1 issue (see
https://github.com/canonical/data-platform-libs/issues/252), or even a
Juju issue.

2) Client CA not recognized as a "new" CA
A unit of etcd might receive the renewed CA first from the client
application before it receives it from the TLS provider. It will add the
new CA to its client truststore already. When the "new" CA is received
in the `certificate_available` event, it will not be recognized as such,
and the unit will therefore not participate in the CA rotation of the
cluster. The CA rotation then gets stuck and never completes.

3) Restarts interfere with each other
If multiple rolling restarts are requested at the same time, rolling ops
will only send one callback per unit. That means, if etcd requests a
restart because of CA rotation and, at the same time, a new client
application CA also requires/requests a restart, one of the restart
requests gets lost. This is known behavior, see
[rolling_ops](https://github.com/canonical/charmed-etcd-operator/blob/3.6/edge/lib/charms/rolling_ops/v0/rollingops.py#L146).

## Solution:
1) Temporary fix for the `secret_changed` event handler in
data-interfaces v1 until issue
https://github.com/canonical/data-platform-libs/issues/252 is resolved.

2) Adjustment to the `is_new_ca_saved_on_all_servers` check in the
`TLSManager`: units with no CA rotation are ignored (in case of client
TLS).

3) Consider the restart request issued when updating the client
mtls-certs a "ca_rotation_restart", because it also loads the truststore
and can continue the ca-rotation workflow if it happens at the same
time.

## Drive-by changes:
In the fashion of #188: add temporary workaround for Pydantic
serialization issues. These can also occur when updating client relation
data because of changes on TLS. This workaround will be removed when
https://github.com/canonical/data-platform-libs/issues/251 is fixed.

## How was this change tested?
- [X] Manually
- [X] Unit tests
- [X] Integration tests

Integration test coverage has been adjusted: Client applications now use
the same TLS provider as etcd to ensure a deployment scenario close to
possible production setup.